commit fix_nsjail.patch...
which currently needs to be applied to your AOSP checkout in order for
the nsjail_build_sandbox mechanism to work on nixos. See inline
comments for why that is so.

We'll try to remove this patch in favor of solving this on the file
system layer (i.e. no symlinks)