sign-disk-image: rewrite for raw files only, but runnable in sandbox
Our original implementation re-used a clever trick to avoid copying
all of the image from the read-only nix store by using a qcow2 image
and it's COW layer, while still being able to use raw files for
production images.

We dropped support for qcow2 and do support raw files only now, so
we do need to copy the whole image out of the nix store before signing
and assume that the image file passed in $1 is writable.

We do, however, gain the ability to mount and modify fat32 partitions
inside the image without root or any daemons involved, so that we
can install a signed EFI binary and keys into the image, even while
runnig in restricted environments. Such as the nix sandbox during
NixOS VM tests.