sign-disk-image: copy image to tempdir before signing
First copying the image by hand and then signing the mutable variant
was a bit cleaner w.r.t. separation of concerns, but it turns out UX
wasn't ideal as users got confused as to which images were signed and
which not.

We aim to improve on that by copying the read-only image from
/nix/store to a temporary file during the script run and move it
to the repositorys top-level directory just before finishing.